[merge_bigmech] Not directly related to this branch, more general commit reorganizing the cvode routines in ./Reactions/Fuego
